Moved logic inside submit evaluation result lib function
